adding btrfs snapshots
This patch enables the use of btrfs subvolumes instead of the rsync
--link-dest feature.

It is based on the original patch by Adrian von Bidder...
http://www.dirvish.org/pipermail/dirvish/2010-December/002357.html

...with the following changes:

- don't use rsync --inplace to prevent fragmentation
  (also see http://www.spinics.net/lists/linux-btrfs/msg40354.html)
- change snapshots to read only after they've been successfully made
  to prevent accidental changing or removal of content
- proper whitespace

+++ dirvish-expire.pl
@@ -169,7 +169,17 @@ for $expire (sort {imsort()} @expires)
 
 	$$Options{'no-run'} and next;
 
-	system("rm -rf $$expire{path}/tree");
+	if ($$Options{btrfs})
+	{
+		# am I not root?
+		if ($< != 0)
+		{
+			system("btrfs property set -ts $$expire{path}/tree ro false > /dev/null");
+		}
+		system("btrfs subvolume delete $$expire{path}/tree > /dev/null");
+	} else {
+		system("rm -rf $$expire{path}/tree");
+	}
 	$$Options{tree} and next;
 
 	system("rm -rf $$expire{path}");
